Use memory map instead of manually specifying addresses
