+++ bittide-instances/bittide-instances.cabal
@@ -228,6 +228,7 @@ library
     Bittide.Instances.Pnr.Switch
     Bittide.Instances.Pnr.Synchronizer
     Bittide.Instances.Tests.AddressableBytesWb
+    Bittide.Instances.Tests.ClockControlWb
     Bittide.Instances.Tests.ElasticBufferWb
     Bittide.Instances.Tests.NestedInterconnect
     Bittide.Instances.Tests.RegisterWb
